+++ test/test_direct_access.py
@@ -2,7 +2,7 @@ import pytest
 from hypothesis import strategies as st
 
 import schemathesis
-from schemathesis.generation.meta import CaseMetadata, GenerationInfo, PhaseInfo
+from schemathesis.generation.meta import CaseMetadata, FuzzingPhaseData, GenerationInfo, PhaseInfo, TestPhase
 from schemathesis.generation.modes import GenerationMode
 from schemathesis.schemas import APIOperation
 
@@ -40,7 +40,17 @@ def test_as_strategy(swagger_20):
     assert isinstance(strategy, st.SearchStrategy)
     assert strategy.example() == operation.Case(
         _meta=CaseMetadata(
-            generation=GenerationInfo(time=0.0, mode=GenerationMode.POSITIVE), components={}, phase=PhaseInfo.fuzzing()
+            generation=GenerationInfo(time=0.0, mode=GenerationMode.POSITIVE),
+            components={},
+            phase=PhaseInfo(
+                name=TestPhase.FUZZING,
+                data=FuzzingPhaseData(
+                    description="",
+                    parameter=None,
+                    parameter_location=None,
+                    location=None,
+                ),
+            ),
         ),
     )