+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.3.17...HEAD) - TBD
 
+### :wrench: Changed
+
+- Show mutation metadata in `negative_data_rejection` check for all testing phases.
+
 ## [4.3.17](https://github.com/schemathesis/schemathesis/compare/v4.3.16...v4.3.17) - 2025-11-02
 
 ### :rocket: Added